requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==4.1.6
django-allauth==0.52.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pypng==0.20220715.0
python3-openid==3.2.0
pytz==2022.7.1
qrcode==7.4.2
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
